requirements.txt
# Core Flask dependencies
Flask>=3.0.0
Werkzeug>=3.0.0

# AI/ML dependencies with flexible versions
openai>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
crewai>=0.120.0
google-generativeai>=0.8.0
google-ai-generativelanguage>=0.6.15

# Database and utilities
python-dotenv>=1.0.0
requests>=2.25.0

# Optional: Remove version pins for problematic packages
# Let pip resolve these automatically
numpy
tenacity
protobuf
packaging
langsmith
langchain-core